+++ server/src/main/java/org/opensearch/action/admin/cluster/node/stats/NodeStats.java
@@ -56,7 +56,7 @@ import org.opensearch.monitor.jvm.JvmStats;
 import org.opensearch.monitor.os.OsStats;
 import org.opensearch.monitor.process.ProcessStats;
 import org.opensearch.node.AdaptiveSelectionStats;
-import org.opensearch.node.NodesPerformanceStats;
+import org.opensearch.node.NodesResourceUsageStats;
 import org.opensearch.script.ScriptCacheStats;
 import org.opensearch.script.ScriptStats;
 import org.opensearch.search.backpressure.stats.SearchBackpressureStats;
@@ -144,7 +144,7 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
     private SearchPipelineStats searchPipelineStats;
 
     @Nullable
-    private NodesPerformanceStats nodesPerformanceStats;
+    private NodesResourceUsageStats resourceUsageStats;
 
     public NodeStats(StreamInput in) throws IOException {
         super(in);
@@ -203,9 +203,9 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
             searchPipelineStats = null;
         }
         if (in.getVersion().onOrAfter(Version.V_3_0_0)) { // make it 2.12 when we backport
-            nodesPerformanceStats = in.readOptionalWriteable(NodesPerformanceStats::new);
+            resourceUsageStats = in.readOptionalWriteable(NodesResourceUsageStats::new);
         } else {
-            nodesPerformanceStats = null;
+            resourceUsageStats = null;
         }
     }
 
@@ -225,7 +225,7 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
         @Nullable DiscoveryStats discoveryStats,
         @Nullable IngestStats ingestStats,
         @Nullable AdaptiveSelectionStats adaptiveSelectionStats,
-        @Nullable NodesPerformanceStats nodesPerformanceStats,
+        @Nullable NodesResourceUsageStats resourceUsageStats,
         @Nullable ScriptCacheStats scriptCacheStats,
         @Nullable IndexingPressureStats indexingPressureStats,
         @Nullable ShardIndexingPressureStats shardIndexingPressureStats,
@@ -251,7 +251,7 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
         this.discoveryStats = discoveryStats;
         this.ingestStats = ingestStats;
         this.adaptiveSelectionStats = adaptiveSelectionStats;
-        this.nodesPerformanceStats = nodesPerformanceStats;
+        this.resourceUsageStats = resourceUsageStats;
         this.scriptCacheStats = scriptCacheStats;
         this.indexingPressureStats = indexingPressureStats;
         this.shardIndexingPressureStats = shardIndexingPressureStats;
@@ -356,8 +356,8 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
     }
 
     @Nullable
-    public NodesPerformanceStats getNodesPerformanceStats() {
-        return nodesPerformanceStats;
+    public NodesResourceUsageStats getResourceUsageStats() {
+        return resourceUsageStats;
     }
 
     @Nullable
@@ -447,7 +447,7 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
             out.writeOptionalWriteable(searchPipelineStats);
         }
         if (out.getVersion().onOrAfter(Version.V_3_0_0)) { // make it 2.12 when we backport
-            out.writeOptionalWriteable(nodesPerformanceStats);
+            out.writeOptionalWriteable(resourceUsageStats);
         }
     }
 
@@ -539,8 +539,8 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
         if (getSearchPipelineStats() != null) {
             getSearchPipelineStats().toXContent(builder, params);
         }
-        if (getNodesPerformanceStats() != null) {
-            getNodesPerformanceStats().toXContent(builder, params);
+        if (getResourceUsageStats() != null) {
+            getResourceUsageStats().toXContent(builder, params);
         }
         return builder;
     }