+++ server/src/test/java/org/opensearch/node/ResourceUsageCollectorServiceTests.java
@@ -18,9 +18,9 @@ import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.core.common.transport.TransportAddress;
-import org.opensearch.ratelimiting.tracker.NodePerformanceTracker;
-import org.opensearch.ratelimiting.tracker.PerformanceTrackerSettings;
-import org.opensearch.test.OpenSearchTestCase;
+import org.opensearch.node.resource.tracker.NodeResourceUsageTracker;
+import org.opensearch.node.resource.tracker.ResourceTrackerSettings;
+import org.opensearch.test.OpenSearchSingleNodeTestCase;
 import org.opensearch.threadpool.TestThreadPool;
 import org.opensearch.threadpool.ThreadPool;
 import org.junit.After;
@@ -35,33 +35,33 @@ import static org.opensearch.test.ClusterServiceUtils.createClusterService;
 import static org.hamcrest.Matchers.greaterThan;
 
 /**
- * Tests for PerformanceCollectorService where we test collect method, get method and whether schedulers
+ * Tests for ResourceUsageCollectorService where we test collect method, get method and whether schedulers
  * are working as expected
  */
-public class PerformanceCollectorServiceTests extends OpenSearchTestCase {
+public class ResourceUsageCollectorServiceTests extends OpenSearchSingleNodeTestCase {
 
     private ClusterService clusterService;
-    private PerformanceCollectorService collector;
+    private ResourceUsageCollectorService collector;
     private ThreadPool threadpool;
-    NodePerformanceTracker tracker;
+    NodeResourceUsageTracker tracker;
 
     @Before
     public void setUp() throws Exception {
         super.setUp();
 
-        threadpool = new TestThreadPool("performance_collector_tests");
+        threadpool = new TestThreadPool("resource_usage_collector_tests");
 
         clusterService = createClusterService(threadpool);
 
         Settings settings = Settings.builder()
-            .put(PerformanceTrackerSettings.GLOBAL_JVM_USAGE_AC_WINDOW_DURATION_SETTING.getKey(), new TimeValue(500, TimeUnit.MILLISECONDS))
+            .put(ResourceTrackerSettings.GLOBAL_JVM_USAGE_AC_WINDOW_DURATION_SETTING.getKey(), new TimeValue(500, TimeUnit.MILLISECONDS))
             .build();
-        tracker = new NodePerformanceTracker(
+        tracker = new NodeResourceUsageTracker(
             threadpool,
             settings,
             new ClusterSettings(settings, ClusterSettings.BUILT_IN_CLUSTER_SETTINGS)
         );
-        collector = new PerformanceCollectorService(tracker, clusterService, threadpool);
+        collector = new ResourceUsageCollectorService(tracker, clusterService, threadpool);
         tracker.start();
         collector.start();
     }
@@ -77,28 +77,28 @@ public class PerformanceCollectorServiceTests extends OpenSearchTestCase {
         tracker.close();
     }
 
-    public void testNodePerformanceStats() {
-        collector.collectNodePerfStatistics("node1", System.currentTimeMillis(), 97, 99);
-        Map<String, NodePerformanceStats> nodeStats = collector.getAllNodeStatistics();
+    public void testResourceUsageStats() {
+        collector.collectNodeResourceUsageStats("node1", System.currentTimeMillis(), 97, 99);
+        Map<String, NodeResourceUsageStats> nodeStats = collector.getAllNodeStatistics();
         assertTrue(nodeStats.containsKey("node1"));
         assertEquals(99.0, nodeStats.get("node1").cpuUtilizationPercent, 0.0);
         assertEquals(97.0, nodeStats.get("node1").memoryUtilizationPercent, 0.0);
 
-        Optional<NodePerformanceStats> nodePerformanceStatistics = collector.getNodeStatistics("node1");
+        Optional<NodeResourceUsageStats> nodeResourceUsageStatsOptional = collector.getNodeStatistics("node1");
 
-        assertNotNull(nodePerformanceStatistics.get());
-        assertEquals(99.0, nodePerformanceStatistics.get().cpuUtilizationPercent, 0.0);
-        assertEquals(97.0, nodePerformanceStatistics.get().memoryUtilizationPercent, 0.0);
+        assertNotNull(nodeResourceUsageStatsOptional.get());
+        assertEquals(99.0, nodeResourceUsageStatsOptional.get().cpuUtilizationPercent, 0.0);
+        assertEquals(97.0, nodeResourceUsageStatsOptional.get().memoryUtilizationPercent, 0.0);
 
-        nodePerformanceStatistics = collector.getNodeStatistics("node2");
-        assertTrue(nodePerformanceStatistics.isEmpty());
+        nodeResourceUsageStatsOptional = collector.getNodeStatistics("node2");
+        assertTrue(nodeResourceUsageStatsOptional.isEmpty());
     }
 
     public void testScheduler() throws Exception {
         /**
          * Wait for cluster state to be ready so that localNode().getId() is ready and we add the values to the map
          */
-        assertBusy(() -> assertTrue(collector.getNodeStatistics(clusterService.localNode().getId()).isPresent()));
+        assertBusy(() -> assertTrue(collector.getNodeStatistics(clusterService.localNode().getId()).isPresent()), 1, TimeUnit.MINUTES);
         assertTrue(collector.getNodeStatistics(clusterService.localNode().getId()).isPresent());
         /**
          * Wait for memory utilization to be reported greater than 0
@@ -131,9 +131,9 @@ public class PerformanceCollectorServiceTests extends OpenSearchTestCase {
             }
             for (int i = 0; i < randomIntBetween(100, 200); i++) {
                 if (randomBoolean()) {
-                    collector.removeNodePerfStatistics(randomFrom(nodes));
+                    collector.removeNodeResourceUsageStats(randomFrom(nodes));
                 }
-                collector.collectNodePerfStatistics(
+                collector.collectNodeResourceUsageStats(
                     randomFrom(nodes),
                     System.currentTimeMillis(),
                     randomIntBetween(1, 100),
@@ -157,7 +157,7 @@ public class PerformanceCollectorServiceTests extends OpenSearchTestCase {
         t3.join();
         t4.join();
 
-        final Map<String, NodePerformanceStats> nodeStats = collector.getAllNodeStatistics();
+        final Map<String, NodeResourceUsageStats> nodeStats = collector.getAllNodeStatistics();
         for (String nodeId : nodes) {
             if (nodeStats.containsKey(nodeId)) {
                 assertThat(nodeStats.get(nodeId).memoryUtilizationPercent, greaterThan(0.0));
@@ -167,8 +167,8 @@ public class PerformanceCollectorServiceTests extends OpenSearchTestCase {
     }
 
     public void testNodeRemoval() {
-        collector.collectNodePerfStatistics("node1", System.currentTimeMillis(), randomIntBetween(1, 100), randomIntBetween(1, 100));
-        collector.collectNodePerfStatistics("node2", System.currentTimeMillis(), randomIntBetween(1, 100), randomIntBetween(1, 100));
+        collector.collectNodeResourceUsageStats("node1", System.currentTimeMillis(), randomIntBetween(1, 100), randomIntBetween(1, 100));
+        collector.collectNodeResourceUsageStats("node2", System.currentTimeMillis(), randomIntBetween(1, 100), randomIntBetween(1, 100));
 
         ClusterState previousState = ClusterState.builder(new ClusterName("cluster"))
             .nodes(
@@ -183,7 +183,7 @@ public class PerformanceCollectorServiceTests extends OpenSearchTestCase {
         ClusterChangedEvent event = new ClusterChangedEvent("test", newState, previousState);
 
         collector.clusterChanged(event);
-        final Map<String, NodePerformanceStats> nodeStats = collector.getAllNodeStatistics();
+        final Map<String, NodeResourceUsageStats> nodeStats = collector.getAllNodeStatistics();
         assertTrue(nodeStats.containsKey("node1"));
         assertFalse(nodeStats.containsKey("node2"));
     }